Use localhost for test

+++ tests/run.sh
@@ -1,7 +1,19 @@
+if [ ! -d www ]; then
+    echo "must be launch from top dir"
+    exit 1
+fi
+
 #make sure ionic serve isn't running
 kill -9 $(lsof -n -ti4TCP:8100)
 
 #/usr/lib/node_modules/protractor/bin/webdriver-manager update --ignore_ssl
+
+
+echo use localhost preprod
+cat <<EOF > www/js/url.js
+var backend_url = 'http://localhost:8080/';
+EOF
+
 
 echo starting ionic serve...
 screen -d -m -L ionic serve --nolivereload --nobrowser --address localhost
@@ -21,5 +33,8 @@ echo ionic serve started
 echo stoping ionic...
 kill -9 $(lsof -n -ti4TCP:8100)
 echo stopped.
+
+echo revert url.js
+git checkout www/js/url.js
 
 echo done
